requirements.txt
# requirements.txt

frame==0.3.1
image==1.5.33
label==0.1.1
pillow==10.3.0
pytoimage==0.0.1
tk==0.1.0
